+++ src/components/files/filesTypes.ts
@@ -1,6 +1,6 @@
 import { DriveItem } from "@microsoft/microsoft-graph-types-beta";
 import { IArchiveClientProgress, IContainer } from "../../common/types";
-import { IJobProgress } from "../../services/backendApi";
+import type { IJobProgress } from "../../services/downloadArchiveApi";
 
 /**
  * Files 入口组件属性。
