+++ src/components/permissions/hooks/usePermissionDraft.ts
@@ -38,10 +38,15 @@ const areEntriesByTabEqual = (
  * - `originalEntriesByTab` 保存本地维护的确认基线，成功写回后会前移到最新快照。
  * - `draftEntriesByTab` 保存当前编辑过程中的变化，只影响弹窗内的临时编辑体验。
  *
- * 需要 `original` 快照的原因：
- * - Close 要丢弃未提交修改，必须知道“恢复到哪里”。
- * - Apply 成功后要清空脏状态，也必须刷新新的“已确认基线”。
- * - 当容器切换或后端重新加载成功时，也需要把最新服务端结果同步成新的基线。
+ * 需要 original 拷贝的理由:弹窗内部的编辑、确认、撤销逻辑形成自洽的闭环
+ *
+ * - 假设没有，只保留draft和initial，如果后台数据更新了，initial也变了，用户点击reset
+ * 发现数据变了，可能会困惑，使用 original 进行快照符合 Reset 的预期语义。
+ *
+ * - 假设 Apply 提交后，网络延迟，虽然成功了，此时别人立刻又改了权限，如果直接对照 initial，就会发现
+ * 又变了，又有未保存的更改，用户可能也会疑惑：究竟有没有保存？
+ *
+ * - initial 依靠父组件的 re-render 来更新，apply 后会没来得及更新至最新值，用户也会看到还有未保存更改
  */
 export const usePermissionDraft = (
   initialEntriesByTab: PermissionEntriesByTab,
@@ -113,22 +118,13 @@ export const usePermissionDraft = (
   };
 
   /**
-   * 仅在本地把当前草稿提升成新的确认快照。
+   * 用一份新的权限数据同时更新 original 和 draft 快照，保持两者一致。
    *
-   * 这仍然保留给纯本地场景或单元测试使用；
-   * 真正的容器权限写回完成后，会优先使用 `replaceEntries` 把服务端最新结果同步回来。
-   */
-  const applyDraft = () => {
-    const nextOriginalEntriesByTab = cloneEntriesByTab(draftEntriesByTab);
-    setOriginalEntriesByTab(nextOriginalEntriesByTab);
-    setDraftEntriesByTab(cloneEntriesByTab(nextOriginalEntriesByTab));
-  };
-
-  /**
-   * 用服务端最新确认的权限快照同时重置 original 和 draft。
+   * 具体操作：接收权限列表，将其复制后分别赋值给 original 和 draft，
+   * 以此清除任何本地未保存的编辑痕迹。
    *
-   * 这个方法主要用于两种场景：
-   * 1. Dialog 初次打开后，加载回来的真实容器权限；
+   * 主要用于两种场景：
+   * 1. Dialog 初次打开后，用加载回来的真实容器权限同步状态；
    * 2. Apply 成功后，用后端最新结果覆盖本地草稿，清空脏状态。
    */
   const replaceEntries = (entriesByTab: PermissionEntriesByTab) => {
@@ -148,7 +144,6 @@ export const usePermissionDraft = (
     updateEntryRole,
     removeEntry,
     resetDraft,
-    applyDraft,
     replaceEntries,
   };
 };
